feat: 移除列表
